apply --execute: really execute
This change allows the `apply` command when supplied with the `--execute`
flag to always execute the calculated commands and do a `write_ring` or
`rebalance`. Before the commands are only executed when in non interactive
mode after user confirmation. Rebalancing only happend with user confirmation.
Reasoning: After some month of heavy usage of the tool, the user confirmation
becomes more and more obsolete because one need anyhow check the output and
do a `git checkout` in case of issues. The old behaviour is still the default,
if `--execute` is not supplied.

Addiotional the possible multiline output of a rebalance command is now split
into multiple log lines for better readability (`go-bits/log` will not allow
multiline logs)